few changes

+++ src/main/java/com/keyword/scoreestimation/response/KeyWordScoreResponse.java
@@ -2,17 +2,17 @@ package com.keyword.scoreestimation.response;
 
 public class KeyWordScoreResponse {
 	String keyWord;
-	Integer score;
+	Double score;
 	public String getKeyWord() {
 		return keyWord;
 	}
 	public void setKeyWord(String keyWord) {
 		this.keyWord = keyWord;
 	}
-	public Integer getScore() {
+	public Double getScore() {
 		return score;
 	}
-	public void setScore(Integer score) {
+	public void setScore(Double score) {
 		this.score = score;
 	}
